+++ src/generator/index.ts
@@ -1,5 +1,6 @@
 export * from "./block-generator";
 export * from "./code-generator";
+export * from "./tree-generator";
 export * from "./utils";
 export * from "./blocks/function-call";
 export * from "./blocks/if-block";
