+++ src/generator/tree-generator.ts
@@ -1,27 +1,4 @@
-import { CodeBlock } from "../types";
-
-export interface TreeNode {
-  children: TreeNode[];
-  block: CodeBlock;
-  isNestable: boolean;
-  hierarchy: number;
-}
-
-export interface NestableNode extends TreeNode {
-  isNestable: true;
-}
-
-export interface NonNestableNode extends TreeNode {
-  isNestable: false;
-}
-
-export type CodeNode = NestableNode | NonNestableNode;
-
-export interface FlattenedNode {
-  block: CodeBlock;
-  depth: number;
-  isNestable: boolean;
-}
+import { CodeBlock, CodeNode, FlattenedNode } from "../types";
 
 export const blockToTreeNode = (
   block: CodeBlock,
