+++ src/types/index.ts
@@ -2,3 +2,4 @@ export * from "./blocks";
 export * from "./modules";
 export * from "./common";
 export * from "./generator";
+export * from "./tree";
